+++ pyimouapi/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 
 from .device import ImouChannel, ImouDevice, ImouDeviceManager
 from .exceptions import (